Spec sheet requests
Hopefully this works, I'm not sure how to run it. It has a dictionary of sensor-link pairs (links are to the spec sheets in my drive so it doesn't change with different canvas pages). When a user sends a message beginning with !spec, if one of the keys is the same as the rest of the message, the user will be sent a DM with the link and a confirmation of sending will be posted in the public channel